Critical Equipment
Script is written in a procedural format. To use script, make changes to the first dictionary listed called the critical dictionary. You will add the equipment name as listed on a Battalion's UDL along with their equipment and model number. Next, you need to download the equipment status report from G-Army. Finally, execute the script and it will return the quantity of each piece of critical equipment along with their bumper number.